integrated assets area chart with backend
